Fix backwards compatibility issue for reports
