+++ Core/src/org/sleuthkit/autopsy/report/Bundle.properties-MERGED
@@ -2,3 +2,16 @@ OpenIDE-Module-Name=Report
 DefaultReportConfigurationPanel.infoLabel.text=This report will be configured on the next screen.
 ReportBranding.defaultReportTitle.text=Autopsy Forensic Report
 ReportBranding.defaultReportFooter.text=Powered by Autopsy Open Source Digital Forensics Platform - www.sleuthkit.org
+ReportProgressPanel.pathLabel.text=pathLabel
+ReportProgressPanel.reportLabel.text=reportLabel
+ReportProgressPanel.statusMessageLabel.text=processingLabel
+ReportProgressPanel.separationLabel.text=:
+ReportProgressPanel.progress.queuing=Queuing...
+ReportProgressPanel.initPathLabel.noFile=<html><u>No report file</u></html>
+ReportProgressPanel.start.cancelButton.text=Cancel
+ReportProgressPanel.start.progress.text=Starting report...
+ReportProgressPanel.complete.processLbl.text=Complete
+ReportProgressPanel.complete.processLb2.text=Completed with error
+ReportProgressPanel.complete.cancelButton.text=Complete
+ReportProgressPanel.cancel.cancelButton.toolTipText=Canceled
+ReportProgressPanel.cancel.procLbl.text=Canceled
