+++ Core/src/org/sleuthkit/autopsy/report/infrastructure/Bundle.properties
@@ -6,10 +6,6 @@ ReportWizardFileOptionsVisualPanel.selectAllButton.text=Select All
 ReportWizardPortableCaseOptionsVisualPanel.errorLabel.text=Windows only
 ReportWizardPortableCaseOptionsVisualPanel.compressCheckbox.toolTipText=
 ReportWizardPortableCaseOptionsVisualPanel.compressCheckbox.text=Package case into an archive:
-ReportProgressDialog.reportLabel.text=reportLabel
-ReportProgressDialog.statusMessageLabel.text=processingLabel
-ReportProgressDialog.separationLabel.text=:
-ReportProgressDialog.pathLabel.text=pathLabel
 PortableCaseTagsListPanel.jAllTagsCheckBox.text=All Tagged Results
 PortableCaseTagsListPanel.deselectButton.text=Deselect All
 PortableCaseTagsListPanel.selectButton.text=Select All
@@ -151,15 +147,6 @@ ReportGenerator.artTableColHdr.localPath=Local path
 ReportGenerator.artTableColHdr.remotePath=Remote path
 ReportGenerator.errors.reportErrorTitle=Error generating report
 ReportGenerator.errors.reportErrorText=Error generating report: 
-ReportProgressDialog.progress.queuing=Queuing...
-ReportProgressDialog.initPathLabel.noFile=<html><u>No report file</u></html>
-ReportProgressDialog.start.cancelButton.text=Cancel
-ReportProgressDialog.start.progress.text=Starting report...
-ReportProgressDialog.complete.processLbl.text=Complete
-ReportProgressDialog.complete.processLb2.text=Completed with error
-ReportProgressDialog.complete.cancelButton.text=Complete
-ReportProgressDialog.cancel.cancelButton.toolTipText=Canceled
-ReportProgressDialog.cancel.procLbl.text=Canceled
 ReportVisualPanel1.getName.text=Select and Configure Report Modules
 ReportVisualPanel2.getName.text=Configure Report
 ReportWizardAction.actionName.text=Generate Report
