+++ Core/src/org/sleuthkit/autopsy/report/infrastructure/Bundle.properties-MERGED
@@ -9,16 +9,18 @@ PortableCaseTagsListPanel.error.noOpenCase=There is no case open
 ReportGenerator.artTableColHdr.comment=Comment
 ReportGenerator.errList.noOpenCase=No open case available.
 ReportGenerator.tagTable.header.userName=User Name
+ReportProgressIndicator.cancelledMessage=Report generation cancelled
+ReportProgressIndicator.completedMessage=Report generation completed
+ReportProgressIndicator.completedWithErrorsMessage=Report generation completed with errors
+ReportProgressIndicator.startMessage=Report generation started
+ReportProgressIndicator.switchToDeterminateMessage=Report generation progress switched to determinate
+ReportProgressIndicator.switchToIndeterminateMessage=Report generation progress switched to indeterminate
 ReportWizardFileOptionsVisualPanel.jLabel1.text=Select items to include in File Report:
 ReportWizardFileOptionsVisualPanel.deselectAllButton.text=Deselect All
 ReportWizardFileOptionsVisualPanel.selectAllButton.text=Select All
 ReportWizardPortableCaseOptionsVisualPanel.errorLabel.text=Windows only
 ReportWizardPortableCaseOptionsVisualPanel.compressCheckbox.toolTipText=
 ReportWizardPortableCaseOptionsVisualPanel.compressCheckbox.text=Package case into an archive:
-ReportProgressDialog.reportLabel.text=reportLabel
-ReportProgressDialog.statusMessageLabel.text=processingLabel
-ReportProgressDialog.separationLabel.text=:
-ReportProgressDialog.pathLabel.text=pathLabel
 PortableCaseTagsListPanel.jAllTagsCheckBox.text=All Tagged Results
 PortableCaseTagsListPanel.deselectButton.text=Deselect All
 PortableCaseTagsListPanel.selectButton.text=Select All
@@ -160,15 +162,6 @@ ReportGenerator.artTableColHdr.localPath=Local path
 ReportGenerator.artTableColHdr.remotePath=Remote path
 ReportGenerator.errors.reportErrorTitle=Error generating report
 ReportGenerator.errors.reportErrorText=Error generating report: 
-ReportProgressDialog.progress.queuing=Queuing...
-ReportProgressDialog.initPathLabel.noFile=<html><u>No report file</u></html>
-ReportProgressDialog.start.cancelButton.text=Cancel
-ReportProgressDialog.start.progress.text=Starting report...
-ReportProgressDialog.complete.processLbl.text=Complete
-ReportProgressDialog.complete.processLb2.text=Completed with error
-ReportProgressDialog.complete.cancelButton.text=Complete
-ReportProgressDialog.cancel.cancelButton.toolTipText=Canceled
-ReportProgressDialog.cancel.procLbl.text=Canceled
 ReportVisualPanel1.getName.text=Select and Configure Report Modules
 ReportVisualPanel2.getName.text=Configure Report
 ReportWizardAction.actionName.text=Generate Report
