+++ Core/src/org/sleuthkit/autopsy/report/infrastructure/TableReportGenerator.java
@@ -46,8 +46,8 @@ import org.sleuthkit.autopsy.casemodule.services.TagsManager;
 import org.sleuthkit.autopsy.coreutils.ImageUtils;
 import org.sleuthkit.autopsy.coreutils.Logger;
 import org.sleuthkit.autopsy.datamodel.ContentUtils;
-import org.sleuthkit.autopsy.report.ReportProgressPanel;
 import static org.sleuthkit.autopsy.casemodule.services.TagsManager.getNotableTagLabel;
+import org.sleuthkit.autopsy.report.ReportProgressPanel;
 import org.sleuthkit.datamodel.AbstractFile;
 import org.sleuthkit.datamodel.Account;
 import org.sleuthkit.datamodel.BlackboardArtifact;
@@ -60,6 +60,7 @@ import org.sleuthkit.datamodel.SleuthkitCase;
 import org.sleuthkit.datamodel.TagName;
 import org.sleuthkit.datamodel.TskCoreException;
 import org.sleuthkit.datamodel.TskData;
+import org.sleuthkit.autopsy.report.ReportProgressPanel;
 
 class TableReportGenerator {
 
@@ -80,10 +81,10 @@ class TableReportGenerator {
         this.progressPanel = progressPanel;
         this.tableReport = tableReport;
         this.columnHeaderMap = new HashMap<>();
-        errorList = new ArrayList<>();        
+        errorList = new ArrayList<>();
         this.settings = settings;
     }
-    
+
     private void getAllExistingTags() throws NoCurrentCaseException, TskCoreException {
         List<String> tagNames = new ArrayList<>();
 
@@ -98,6 +99,7 @@ class TableReportGenerator {
         tagNamesFilter = new HashSet<>(tagNames);
     }
 
+    @SuppressWarnings("deprecation")
     private void getAllExistingArtiactTypes() throws NoCurrentCaseException, TskCoreException {
         // get all possible artifact types
         ArrayList<BlackboardArtifact.Type> doNotReport = new ArrayList<>();
@@ -116,7 +118,7 @@ class TableReportGenerator {
 
         progressPanel.start();
         progressPanel.updateStatusLabel(NbBundle.getMessage(this.getClass(), "ReportGenerator.progress.readingTagsArtifacts.text"));
-        
+
         if (settings.useStoredTagsAndArtifactsLists()) {
             // Get the artifact types selected by the user.
             artifactTypes = settings.getArtifactSelections();
@@ -130,7 +132,7 @@ class TableReportGenerator {
                 if (settings.getSelectedReportOption() == TableReportSettings.TableReportOption.ALL_TAGGED_RESULTS) {
                     getAllExistingTags();
                 }
-                
+
                 // get all possible artifact types
                 getAllExistingArtiactTypes();
             } catch (NoCurrentCaseException | TskCoreException ex) {
@@ -139,11 +141,11 @@ class TableReportGenerator {
                 return;
             }
         }
-        
+
         // Start the progress indicators for each active TableReportModule.
         progressPanel.setIndeterminate(false);
         progressPanel.setMaximumProgress(this.artifactTypes.size() + 2); // +2 for content and blackboard artifact tags
-        
+
         // report on the blackboard results
         if (progressPanel.getStatus() != ReportProgressPanel.ReportStatus.CANCELED) {
             makeBlackboardArtifactTables();
