+++ Core/src/org/sleuthkit/autopsy/report/modules/bodyfile/BodyFileReport.java
@@ -35,9 +35,9 @@ import org.sleuthkit.autopsy.casemodule.Case;
 import org.sleuthkit.autopsy.casemodule.NoCurrentCaseException;
 import org.sleuthkit.autopsy.coreutils.Logger;
 import org.sleuthkit.autopsy.ingest.IngestManager;
-import org.sleuthkit.autopsy.report.ReportProgressPanel;
 import org.sleuthkit.autopsy.report.ReportProgressPanel.ReportStatus;
 import org.sleuthkit.datamodel.*;
+import org.sleuthkit.autopsy.report.ReportProgressPanel;
 
 /**
  * BodyFileReport generates a report in the body file format specified on The
